Added secret and removalPolicy
Create the Google Auth client secret in the stack and update its value in the console (This may need to be changed to be manually create secret, add value and import into stack at creation depending on how it flows). Added a removalPolicy prop for the cognito userpool as it is retained on stack destruction by default. The stack deploy in this current state.